moved go_live_settings_dialog.dart to widgets
